// src/extensions/searchBar/components/model/IPersonItem.ts
export interface IPersonItem {
    JobTitle: string;
    DisplayName: string;
    FirstName: string;
    LastName: string;
    Department: string;
    Office: string;
    WorkPhone: string;
    PictureUrl: string;
    ProfileUrl: string;
}

export interface Hyperlink {
    Url: string;
    Description: string;
}